Refactoring for modularization
